requirements.txt
# Core
streamlit==1.29.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1

# Binance API
python-binance==1.0.19
websocket-client==1.7.0
ccxt==4.1.56

# Data Processing
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
scipy==1.11.4

# Visualization
plotly==5.18.0
altair==5.2.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
streamlit-authenticator==0.2.3

# AI/ML
scikit-learn==1.3.2
tensorflow==2.15.0
prophet==1.1.5
statsmodels==0.14.1
xgboost==2.0.3
lightgbm==4.1.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
pytz==2023.3.post1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.1